Allow choosing which remote to query for PRs
Useful in cases where multiple remotes are present locally. In that case,
currently a random (?) repo is chosen and if that's not the one being used
to file PRs, then it doesn't work at all.

This will make the default choice be the `origin` remote, but it can be
overridden with `-r` or `--remote`.